[MLG-1] -Moved to core.
