+++ weather-forecast-service-ingestor/src/main/java/com/training/weather/ingestor/core/service/WeatherDataSource.java
@@ -1,7 +1,7 @@
-package com.training.weather.ingestor.infrastructure.service;
+package com.training.weather.ingestor.core.service;
 
 import com.training.weather.ingestor.core.model.Coordinates;
-import com.training.weather.ingestor.core.model.OpenWeatherMapResponse;
+import com.training.weather.ingestor.infrastructure.entity.redis.OpenWeatherMapResponse;
 
 public interface WeatherDataSource {
   OpenWeatherMapResponse getForecasts(Coordinates coordinates);
